requirements.txt
django
djangorestframework
stripe
decouple